route delete terminée

+++ backend/controllers/Book.js
@@ -1,4 +1,5 @@
 const Book = require('../models/Book');
+const fs = require ('fs');
 
 exports.createBook = (req, res) => {
   try {
@@ -49,9 +50,20 @@ exports.modifyBook = (req, res) => {
 };
 
 exports.deleteBook = (req, res) => {
-    Book.deleteOne({_id: req.params.id})
-    .then(()=> res.status(200).json({ message: 'Livre supprimé !'}))
-    .catch(error => res.status(404).json({ error }));
+    Book.findOne({_id: req.params.id})
+    .then(book => {
+      if(book.userId != req.auth.userId){
+        res.status(401).json({message : 'non-autorisé' });
+      } else {
+        const filename = book.imageUrl.split('/images/')[1];
+        fs.unlink(`images/${filename}`, () => {
+          Book.deleteOne({_id: req.params.id})
+          .then(() => {res.status(200).json({message : 'Livre supprimé'})})
+          .catch(error => res.status(401).json({ error }));
+        })
+      }
+    })
+    .catch(error => res.status(500).json({ error }));
   };
 
 exports.getOneBook = (req, res) =>{
